+++ tests_pytest/common_tests/unit_tests/core/common/mixed_precision/test_filter_candidates_for_mixed_precision.py
