+++ tests_pytest/keras_tests/unit_tests/trainable_infrastructure/test_linear_annealing.py
